feat: add verify snapshot testing for `InterfaceStubGeneratorTests`
