+++ Refit.Tests/InterfaceStubGenerator.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.IO;
-using System.Linq;
+﻿using System.Collections.Immutable;
 using System.Reflection;
-using System.Threading.Tasks;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
@@ -12,11 +7,9 @@ using Microsoft.CodeAnalysis.Testing;
 
 using Refit.Generator;
 
-using Xunit;
+using VerifyTests.DiffPlex;
 
 using Task = System.Threading.Tasks.Task;
-using VerifyCS = Refit.Tests.CSharpSourceGeneratorVerifier<Refit.Generator.InterfaceStubGenerator>;
-using VerifyCSV2 = Refit.Tests.CSharpIncrementalSourceGeneratorVerifier<Refit.Generator.InterfaceStubGeneratorV2>;
 
 namespace Refit.Tests;
 
@@ -50,6 +43,35 @@ public class InterfaceStubGeneratorTests
             .AddAssemblies(ImmutableArray.Create("System.Web"))
             .AddPackages(ImmutableArray.Create(new PackageIdentity("System.Net.Http", "4.3.4")));
 #endif
+
+        // Initialise Verify
+        DerivePathInfo((file, _, type, method) => new(Path.Combine(Path.GetDirectoryName(file), "_snapshots"), type.Name, method.Name));
+
+        VerifySourceGenerators.Initialize();
+        VerifyDiffPlex.Initialize(OutputType.Compact);
+    }
+
+    public static async Task<VerifyResult> VerifyGenerator(string input)
+    {
+        var assemblies = await ReferenceAssemblies.ResolveAsync(null, default);
+
+        string[] inputs = [input];
+        var compilation = CSharpCompilation.Create(
+            "compilation",
+            inputs.Select(source => CSharpSyntaxTree.ParseText(File.ReadAllText(source))),
+            assemblies.Add(RefitAssembly),
+            new CSharpCompilationOptions(OutputKind.ConsoleApplication)
+        );
+
+#if NET48
+        var generator = new InterfaceStubGenerator();
+#else
+        var generator = new InterfaceStubGeneratorV2();
+#endif
+        var driver = CSharpGeneratorDriver.Create(generator);
+        var ranDriver = driver.RunGenerators(compilation);
+
+        return await Verify(ranDriver).ToTask();
     }
 
     [Fact(Skip = "Generator in test issue")]
@@ -114,845 +136,22 @@ public class InterfaceStubGeneratorTests
     [Fact]
     public async Task NoRefitInterfacesSmokeTest()
     {
-#if NET48
-        var input = File.ReadAllText(IntegrationTestHelper.GetPath("IInterfaceWithoutRefit.cs"));
-#else
-        var input = await File.ReadAllTextAsync(
-            IntegrationTestHelper.GetPath("IInterfaceWithoutRefit.cs")
-        );
-#endif
-
-        await new VerifyCS.Test
-        {
-            ReferenceAssemblies = ReferenceAssemblies,
-            TestState = { AdditionalReferences = { RefitAssembly }, Sources = { input }, },
-        }.RunAsync();
-
-        await new VerifyCSV2.Test
-        {
-            ReferenceAssemblies = ReferenceAssemblies,
-            TestState = { AdditionalReferences = { RefitAssembly }, Sources = { input }, },
-        }.RunAsync();
+        var path = IntegrationTestHelper.GetPath("IInterfaceWithoutRefit.cs");
+        await VerifyGenerator(path);
     }
 
     [Fact]
     public async Task FindInterfacesSmokeTest()
     {
-#if NET48
-        var input = File.ReadAllText(IntegrationTestHelper.GetPath("GitHubApi.cs"));
-#else
-        var input = await File.ReadAllTextAsync(IntegrationTestHelper.GetPath("GitHubApi.cs"));
-#endif
-
-        var output1 =
-            @"
-#pragma warning disable
-namespace RefitInternalGenerated
-{
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    [global::System.AttributeUsage (global::System.AttributeTargets.Class | global::System.AttributeTargets.Struct | global::System.AttributeTargets.Enum | global::System.AttributeTargets.Constructor | global::System.AttributeTargets.Method | global::System.AttributeTargets.Property | global::System.AttributeTargets.Field | global::System.AttributeTargets.Event | global::System.AttributeTargets.Interface | global::System.AttributeTargets.Delegate)]
-    sealed class PreserveAttribute : global::System.Attribute
-    {
-        //
-        // Fields
-        //
-        public bool AllMembers;
-
-        public bool Conditional;
-    }
-}
-#pragma warning restore
-";
-
-        var output1_5 =
-            @"
-#pragma warning disable
-namespace Refit.Implementation
-{
-
-    /// <inheritdoc />
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.Diagnostics.DebuggerNonUserCode]
-    [global::RefitInternalGenerated.PreserveAttribute]
-    [global::System.Reflection.Obfuscation(Exclude=true)]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    internal static partial class Generated
-    {
-#if NET5_0_OR_GREATER
-        [System.Runtime.CompilerServices.ModuleInitializer]
-        [System.Diagnostics.CodeAnalysis.DynamicDependency(System.Diagnostics.CodeAnalysis.DynamicallyAccessedMemberTypes.All, typeof(global::Refit.Implementation.Generated))]
-        public static void Initialize()
-        {
-        }
-#endif
-    }
-}
-#pragma warning restore
-";
-
-        var output2 = """
-#nullable disable
-#pragma warning disable
-namespace Refit.Implementation
-{
-
-    partial class Generated
-    {
-
-    /// <inheritdoc />
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.Diagnostics.DebuggerNonUserCode]
-    [global::RefitInternalGenerated.PreserveAttribute]
-    [global::System.Reflection.Obfuscation(Exclude=true)]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    partial class RefitTestsIGitHubApi
-        : global::Refit.Tests.IGitHubApi
-    {
-        /// <inheritdoc />
-        public global::System.Net.Http.HttpClient Client { get; }
-        readonly global::Refit.IRequestBuilder requestBuilder;
-
-        /// <inheritdoc />
-        public RefitTestsIGitHubApi(global::System.Net.Http.HttpClient client, global::Refit.IRequestBuilder requestBuilder)
-        {
-            Client = client;
-            this.requestBuilder = requestBuilder;
-        }
-
-
-        private static readonly global::System.Type[] ______typeParameters = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.Tests.User> GetUser(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUser", ______typeParameters );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters0 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public global::System.IObservable<global::Refit.Tests.User> GetUserObservable(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserObservable", ______typeParameters0 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters1 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public global::System.IObservable<global::Refit.Tests.User> GetUserCamelCase(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserCamelCase", ______typeParameters1 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters2 = new global::System.Type[] {typeof(string), typeof(global::System.Threading.CancellationToken) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>> GetOrgMembers(string @orgName, global::System.Threading.CancellationToken @cancellationToken)
-        {
-            var ______arguments = new object[] { @orgName, @cancellationToken };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetOrgMembers", ______typeParameters2 );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters3 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult> FindUsers(string @q)
-        {
-            var ______arguments = new object[] { @q };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("FindUsers", ______typeParameters3 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage> GetIndex()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndex", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        public global::System.IObservable<string> GetIndexObservable()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndexObservable", global::System.Array.Empty<global::System.Type>() );
-
-            return (global::System.IObservable<string>)______func(this.Client, ______arguments);
-        }
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.Tests.User> NothingToSeeHere()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("NothingToSeeHere", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>> NothingToSeeHereWithMetadata()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("NothingToSeeHereWithMetadata", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters4 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>> GetUserWithMetadata(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserWithMetadata", ______typeParameters4 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters5 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public global::System.IObservable<global::Refit.ApiResponse<global::Refit.Tests.User>> GetUserObservableWithMetadata(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserObservableWithMetadata", ______typeParameters5 );
-
-            return (global::System.IObservable<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters6 = new global::System.Type[] {typeof(global::Refit.Tests.User) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.Tests.User> CreateUser(global::Refit.Tests.User @user)
-        {
-            var ______arguments = new object[] { @user };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("CreateUser", ______typeParameters6 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters7 = new global::System.Type[] {typeof(global::Refit.Tests.User) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>> CreateUserWithMetadata(global::Refit.Tests.User @user)
-        {
-            var ______arguments = new object[] { @user };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("CreateUserWithMetadata", ______typeParameters7 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters8 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.Tests.User> global::Refit.Tests.IGitHubApi.GetUser(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUser", ______typeParameters8 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters9 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        global::System.IObservable<global::Refit.Tests.User> global::Refit.Tests.IGitHubApi.GetUserObservable(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserObservable", ______typeParameters9 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters10 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        global::System.IObservable<global::Refit.Tests.User> global::Refit.Tests.IGitHubApi.GetUserCamelCase(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserCamelCase", ______typeParameters10 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters11 = new global::System.Type[] {typeof(string), typeof(global::System.Threading.CancellationToken) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>> global::Refit.Tests.IGitHubApi.GetOrgMembers(string @orgName, global::System.Threading.CancellationToken @cancellationToken)
-        {
-            var ______arguments = new object[] { @orgName, @cancellationToken };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetOrgMembers", ______typeParameters11 );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters12 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult> global::Refit.Tests.IGitHubApi.FindUsers(string @q)
-        {
-            var ______arguments = new object[] { @q };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("FindUsers", ______typeParameters12 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage> global::Refit.Tests.IGitHubApi.GetIndex()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndex", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        global::System.IObservable<string> global::Refit.Tests.IGitHubApi.GetIndexObservable()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndexObservable", global::System.Array.Empty<global::System.Type>() );
-
-            return (global::System.IObservable<string>)______func(this.Client, ______arguments);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.Tests.User> global::Refit.Tests.IGitHubApi.NothingToSeeHere()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("NothingToSeeHere", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>> global::Refit.Tests.IGitHubApi.NothingToSeeHereWithMetadata()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("NothingToSeeHereWithMetadata", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters13 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>> global::Refit.Tests.IGitHubApi.GetUserWithMetadata(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserWithMetadata", ______typeParameters13 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters14 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        global::System.IObservable<global::Refit.ApiResponse<global::Refit.Tests.User>> global::Refit.Tests.IGitHubApi.GetUserObservableWithMetadata(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserObservableWithMetadata", ______typeParameters14 );
-
-            return (global::System.IObservable<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters15 = new global::System.Type[] {typeof(global::Refit.Tests.User) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.Tests.User> global::Refit.Tests.IGitHubApi.CreateUser(global::Refit.Tests.User @user)
-        {
-            var ______arguments = new object[] { @user };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("CreateUser", ______typeParameters15 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters16 = new global::System.Type[] {typeof(global::Refit.Tests.User) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>> global::Refit.Tests.IGitHubApi.CreateUserWithMetadata(global::Refit.Tests.User @user)
-        {
-            var ______arguments = new object[] { @user };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("CreateUserWithMetadata", ______typeParameters16 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.ApiResponse<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-    }
-    }
-}
-
-#pragma warning restore
-
-""";
-        var output3 =
-            @"#nullable disable
-#pragma warning disable
-namespace Refit.Implementation
-{
-
-    partial class Generated
-    {
-
-    /// <inheritdoc />
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.Diagnostics.DebuggerNonUserCode]
-    [global::RefitInternalGenerated.PreserveAttribute]
-    [global::System.Reflection.Obfuscation(Exclude=true)]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    partial class RefitTestsIGitHubApiDisposable
-        : global::Refit.Tests.IGitHubApiDisposable
-    {
-        /// <inheritdoc />
-        public global::System.Net.Http.HttpClient Client { get; }
-        readonly global::Refit.IRequestBuilder requestBuilder;
-
-        /// <inheritdoc />
-        public RefitTestsIGitHubApiDisposable(global::System.Net.Http.HttpClient client, global::Refit.IRequestBuilder requestBuilder)
-        {
-            Client = client;
-            this.requestBuilder = requestBuilder;
-        }
-
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task RefitMethod()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod(""RefitMethod"", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task global::Refit.Tests.IGitHubApiDisposable.RefitMethod()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod(""RefitMethod"", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        void global::System.IDisposable.Dispose()
-        {
-                Client?.Dispose();
-        }
-    }
-    }
-}
-
-#pragma warning restore
-";
-        var output4 = """
-#nullable disable
-#pragma warning disable
-namespace Refit.Implementation
-{
-
-    partial class Generated
-    {
-
-    /// <inheritdoc />
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.Diagnostics.DebuggerNonUserCode]
-    [global::RefitInternalGenerated.PreserveAttribute]
-    [global::System.Reflection.Obfuscation(Exclude=true)]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    partial class RefitTestsTestNestedINestedGitHubApi
-        : global::Refit.Tests.TestNested.INestedGitHubApi
-    {
-        /// <inheritdoc />
-        public global::System.Net.Http.HttpClient Client { get; }
-        readonly global::Refit.IRequestBuilder requestBuilder;
-
-        /// <inheritdoc />
-        public RefitTestsTestNestedINestedGitHubApi(global::System.Net.Http.HttpClient client, global::Refit.IRequestBuilder requestBuilder)
-        {
-            Client = client;
-            this.requestBuilder = requestBuilder;
-        }
-
-
-        private static readonly global::System.Type[] ______typeParameters = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.Tests.User> GetUser(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUser", ______typeParameters );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters0 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public global::System.IObservable<global::Refit.Tests.User> GetUserObservable(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserObservable", ______typeParameters0 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters1 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public global::System.IObservable<global::Refit.Tests.User> GetUserCamelCase(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserCamelCase", ______typeParameters1 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters2 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>> GetOrgMembers(string @orgName)
-        {
-            var ______arguments = new object[] { @orgName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetOrgMembers", ______typeParameters2 );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters3 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult> FindUsers(string @q)
-        {
-            var ______arguments = new object[] { @q };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("FindUsers", ______typeParameters3 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage> GetIndex()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndex", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        public global::System.IObservable<string> GetIndexObservable()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndexObservable", global::System.Array.Empty<global::System.Type>() );
-
-            return (global::System.IObservable<string>)______func(this.Client, ______arguments);
-        }
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task NothingToSeeHere()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("NothingToSeeHere", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters4 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.Tests.User> global::Refit.Tests.TestNested.INestedGitHubApi.GetUser(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUser", ______typeParameters4 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.User>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters5 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        global::System.IObservable<global::Refit.Tests.User> global::Refit.Tests.TestNested.INestedGitHubApi.GetUserObservable(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserObservable", ______typeParameters5 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters6 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        global::System.IObservable<global::Refit.Tests.User> global::Refit.Tests.TestNested.INestedGitHubApi.GetUserCamelCase(string @userName)
-        {
-            var ______arguments = new object[] { @userName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetUserCamelCase", ______typeParameters6 );
-
-            return (global::System.IObservable<global::Refit.Tests.User>)______func(this.Client, ______arguments);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters7 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>> global::Refit.Tests.TestNested.INestedGitHubApi.GetOrgMembers(string @orgName)
-        {
-            var ______arguments = new object[] { @orgName };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetOrgMembers", ______typeParameters7 );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Collections.Generic.List<global::Refit.Tests.User>>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        private static readonly global::System.Type[] ______typeParameters8 = new global::System.Type[] {typeof(string) };
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult> global::Refit.Tests.TestNested.INestedGitHubApi.FindUsers(string @q)
-        {
-            var ______arguments = new object[] { @q };
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("FindUsers", ______typeParameters8 );
-
-            return await ((global::System.Threading.Tasks.Task<global::Refit.Tests.UserSearchResult>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage> global::Refit.Tests.TestNested.INestedGitHubApi.GetIndex()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndex", global::System.Array.Empty<global::System.Type>() );
-
-            return await ((global::System.Threading.Tasks.Task<global::System.Net.Http.HttpResponseMessage>)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        global::System.IObservable<string> global::Refit.Tests.TestNested.INestedGitHubApi.GetIndexObservable()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("GetIndexObservable", global::System.Array.Empty<global::System.Type>() );
-
-            return (global::System.IObservable<string>)______func(this.Client, ______arguments);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task global::Refit.Tests.TestNested.INestedGitHubApi.NothingToSeeHere()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod("NothingToSeeHere", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-    }
-    }
-}
-
-#pragma warning restore
-
-""";
-
-        await new VerifyCS.Test
-        {
-            ReferenceAssemblies = ReferenceAssemblies,
-            TestState =
-            {
-                AdditionalReferences = { RefitAssembly },
-                Sources = { input },
-                GeneratedSources =
-                {
-                    (typeof(InterfaceStubGenerator), "PreserveAttribute.g.cs", output1),
-                    (typeof(InterfaceStubGenerator), "Generated.g.cs", output1_5),
-                    (typeof(InterfaceStubGenerator), "IGitHubApi.g.cs", output2),
-                    (typeof(InterfaceStubGenerator), "IGitHubApiDisposable.g.cs", output3),
-                    (typeof(InterfaceStubGenerator), "INestedGitHubApi.g.cs", output4),
-                },
-            },
-        }.RunAsync();
-
-        await new VerifyCSV2.Test
-        {
-            ReferenceAssemblies = ReferenceAssemblies,
-            TestState =
-            {
-                AdditionalReferences = { RefitAssembly },
-                Sources = { input },
-                GeneratedSources =
-                {
-                    (typeof(InterfaceStubGeneratorV2), "PreserveAttribute.g.cs", output1),
-                    (typeof(InterfaceStubGeneratorV2), "Generated.g.cs", output1_5),
-                    (typeof(InterfaceStubGeneratorV2), "IGitHubApi.g.cs", output2),
-                    (typeof(InterfaceStubGeneratorV2), "IGitHubApiDisposable.g.cs", output3),
-                    (typeof(InterfaceStubGeneratorV2), "INestedGitHubApi.g.cs", output4),
-                },
-            },
-        }.RunAsync();
+        var path = IntegrationTestHelper.GetPath("GitHubApi.cs");
+        await VerifyGenerator(path);
     }
 
     [Fact]
     public async Task GenerateInterfaceStubsWithoutNamespaceSmokeTest()
     {
-#if NET48
-        var input = File.ReadAllText(IntegrationTestHelper.GetPath("IServiceWithoutNamespace.cs"));
-#else
-        var input = await File.ReadAllTextAsync(
-            IntegrationTestHelper.GetPath("IServiceWithoutNamespace.cs")
-        );
-#endif
-        var output1 =
-            @"
-#pragma warning disable
-namespace RefitInternalGenerated
-{
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    [global::System.AttributeUsage (global::System.AttributeTargets.Class | global::System.AttributeTargets.Struct | global::System.AttributeTargets.Enum | global::System.AttributeTargets.Constructor | global::System.AttributeTargets.Method | global::System.AttributeTargets.Property | global::System.AttributeTargets.Field | global::System.AttributeTargets.Event | global::System.AttributeTargets.Interface | global::System.AttributeTargets.Delegate)]
-    sealed class PreserveAttribute : global::System.Attribute
-    {
-        //
-        // Fields
-        //
-        public bool AllMembers;
-
-        public bool Conditional;
-    }
-}
-#pragma warning restore
-";
-        var output1_5 =
-            @"
-#pragma warning disable
-namespace Refit.Implementation
-{
-
-    /// <inheritdoc />
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.Diagnostics.DebuggerNonUserCode]
-    [global::RefitInternalGenerated.PreserveAttribute]
-    [global::System.Reflection.Obfuscation(Exclude=true)]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    internal static partial class Generated
-    {
-#if NET5_0_OR_GREATER
-        [System.Runtime.CompilerServices.ModuleInitializer]
-        [System.Diagnostics.CodeAnalysis.DynamicDependency(System.Diagnostics.CodeAnalysis.DynamicallyAccessedMemberTypes.All, typeof(global::Refit.Implementation.Generated))]
-        public static void Initialize()
-        {
-        }
-#endif
-    }
-}
-#pragma warning restore
-";
-
-        var output2 =
-            @"#nullable disable
-#pragma warning disable
-namespace Refit.Implementation
-{
-
-    partial class Generated
-    {
-
-    /// <inheritdoc />
-    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    [global::System.Diagnostics.DebuggerNonUserCode]
-    [global::RefitInternalGenerated.PreserveAttribute]
-    [global::System.Reflection.Obfuscation(Exclude=true)]
-    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-    partial class IServiceWithoutNamespace
-        : global::IServiceWithoutNamespace
-    {
-        /// <inheritdoc />
-        public global::System.Net.Http.HttpClient Client { get; }
-        readonly global::Refit.IRequestBuilder requestBuilder;
-
-        /// <inheritdoc />
-        public IServiceWithoutNamespace(global::System.Net.Http.HttpClient client, global::Refit.IRequestBuilder requestBuilder)
-        {
-            Client = client;
-            this.requestBuilder = requestBuilder;
-        }
-
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task GetRoot()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod(""GetRoot"", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        public async global::System.Threading.Tasks.Task PostRoot()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod(""PostRoot"", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task global::IServiceWithoutNamespace.GetRoot()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod(""GetRoot"", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-
-        /// <inheritdoc />
-        async global::System.Threading.Tasks.Task global::IServiceWithoutNamespace.PostRoot()
-        {
-            var ______arguments = global::System.Array.Empty<object>();
-            var ______func = requestBuilder.BuildRestResultFuncForMethod(""PostRoot"", global::System.Array.Empty<global::System.Type>() );
-
-            await ((global::System.Threading.Tasks.Task)______func(this.Client, ______arguments)).ConfigureAwait(false);
-        }
-    }
-    }
-}
-
-#pragma warning restore
-";
-
-        await new VerifyCS.Test
-        {
-            ReferenceAssemblies = ReferenceAssemblies,
-            TestState =
-            {
-                AdditionalReferences = { RefitAssembly },
-                Sources = { input },
-                GeneratedSources =
-                {
-                    (typeof(InterfaceStubGenerator), "PreserveAttribute.g.cs", output1),
-                    (typeof(InterfaceStubGenerator), "Generated.g.cs", output1_5),
-                    (typeof(InterfaceStubGenerator), "IServiceWithoutNamespace.g.cs", output2),
-                },
-            },
-        }.RunAsync();
-
-        await new VerifyCSV2.Test
-        {
-            ReferenceAssemblies = ReferenceAssemblies,
-            TestState =
-            {
-                AdditionalReferences = { RefitAssembly },
-                Sources = { input },
-                GeneratedSources =
-                {
-                    (typeof(InterfaceStubGeneratorV2), "PreserveAttribute.g.cs", output1),
-                    (typeof(InterfaceStubGeneratorV2), "Generated.g.cs", output1_5),
-                    (typeof(InterfaceStubGeneratorV2), "IServiceWithoutNamespace.g.cs", output2),
-                },
-            },
-        }.RunAsync();
+        var path = IntegrationTestHelper.GetPath("IServiceWithoutNamespace.cs");
+        await VerifyGenerator(path);
     }
 }
