+++ Refit.Tests/_snapshots/InterfaceStubGeneratorTests.GenerateInterfaceStubsWithoutNamespaceSmokeTest#PreserveAttribute.g.verified.cs
@@ -1,0 +1,19 @@
+﻿//HintName: PreserveAttribute.g.cs
+
+#pragma warning disable
+namespace RefitInternalGenerated
+{
+    [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+    [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
+    [global::System.AttributeUsage (global::System.AttributeTargets.Class | global::System.AttributeTargets.Struct | global::System.AttributeTargets.Enum | global::System.AttributeTargets.Constructor | global::System.AttributeTargets.Method | global::System.AttributeTargets.Property | global::System.AttributeTargets.Field | global::System.AttributeTargets.Event | global::System.AttributeTargets.Interface | global::System.AttributeTargets.Delegate)]
+    sealed class PreserveAttribute : global::System.Attribute
+    {
+        //
+        // Fields
+        //
+        public bool AllMembers;
+
+        public bool Conditional;
+    }
+}
+#pragma warning restore
